__int64 was an accomodation for old versions of MSVC, is not needed now and makes life tricky when compiling for Android -- long long and unsigned long long have been part of the language definition for a very long time.

+++ Core/zoolib/AnyBase.h
@@ -497,7 +497,7 @@ private:
 		short fAsShort;
 		int fAsInt;
 		long fAsLong;
-		__int64 fAsLongLong;
+		long long fAsLongLong;
 		float fAsFloat;
 		double fAsDouble;
 		} fPayload;
